requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
mcp>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
pydantic>=2.5.0
asyncpg>=0.29.0